% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements.R
\name{ddl_measurements}
\alias{ddl_measurements}
\title{Haal metingen op voor één locatie (via indexcode)}
\usage{
ddl_measurements(
  locatie_naam = "Hoek van Holland",
  start_date,
  end_date,
  grootheid = "WATHTE",
  groepering = "",
  procestype = "meting",
  hoedanigheid = "NAP",
  convert = TRUE
)
}
\arguments{
\item{start_date, }{end_date Date/POSIXct of ISO-8601 character}

\item{grootheid}{Character filter; default "WATHTE"}

\item{groepering}{Character filter; default "NVT"}

\item{hoedanigheid}{Character filter; default "NAP"}

\item{convert}{Logical; TRUE = data.frame}

\item{index_code}{Character, bv. "HOEKVHLD"}
}
\description{
Haal metingen op voor één locatie (via indexcode)
}
