% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_ddlpy.R
\name{install_ddlpy}
\alias{install_ddlpy}
\title{Installeer de Python package 'ddlpy' in de actieve reticulate-omgeving}
\usage{
install_ddlpy(
  envname = NULL,
  extras = c("netcdf", "examples"),
  method = "auto"
)
}
\arguments{
\item{envname}{Optionele naam van een conda- of virtualenv-omgeving.}

\item{extras}{Vector van extra componenten, standaard c("netcdf","examples").}

\item{method}{Installatiemethode ("auto", "virtualenv", "conda").}
}
\value{
Onzichtbaar `NULL`.
}
\description{
Deze functie gebruikt `reticulate::py_install()` om `rws-ddlpy` te installeren.
Eventueel kun je extra's zoals 'netcdf' en 'examples' meegeven.
}
\examples{
\dontrun{
install_ddlpy()
}
}
